+++ Julia_Fractal_Application/Julia_Fractal_Application/sources/GenerateFractal.cs
@@ -11,7 +11,7 @@ namespace Julia_Fractal_Application.sources
 	class GenerateFractal
 	{
 		[DllImport("D:/Documents/sztudyja/Studies-projects/JuliaFractal_masm/Julia_Fractal_Application/x64/Debug/CPP.dll")]
-		private static extern int colorBase(double cReal, double cImg, int X, int Y, int x, int y);
+		private static extern int colorBase(double cReal, double cImg, double X, double Y, double x, double y);
 
 		public static Bitmap run(double cReal, double cImg)
         {
